% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_plot.R
\name{add_plot_attributes}
\alias{add_plot_attributes}
\title{Add information stored as attributes on plot}
\usage{
add_plot_attributes(x)
}
\arguments{
\item{x}{An object.}
}
\description{
Add information stored as attributes on plot
}
