% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_bluebrown.R
\name{bluebrown_colors}
\alias{bluebrown_colors}
\title{Extract blue-brown colors as hex codes}
\usage{
bluebrown_colors(...)
}
\arguments{
\item{...}{Character names of colors.}
}
\value{
A character vector with color-codes.
}
\description{
Can be used to get the hex code of specific colors from the blue-brown color palette. Use \code{bluebrown_colors()} to see all available color.
}
\examples{
bluebrown_colors()

bluebrown_colors("blue", "brown")
}
