% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_plot.R, R/plot.estimateContrasts.R,
%   R/plot.hdi.R, R/plot.p_direction.R, R/plot.rope.R
\name{data_plot}
\alias{data_plot}
\alias{data_plot.estimateContrasts}
\alias{plot.estimateContrasts}
\alias{data_plot.hdi}
\alias{plot.hdi}
\alias{data_plot.p_direction}
\alias{plot.p_direction}
\alias{data_plot.rope}
\alias{plot.rope}
\title{Prepare an object for plotting}
\usage{
data_plot(x, data = NULL, ...)

\method{data_plot}{estimateContrasts}(x, data = NULL, ...)

\method{plot}{estimateContrasts}(x, data = NULL, ...)

\method{data_plot}{hdi}(x, data = NULL, ...)

\method{plot}{hdi}(x, data = NULL, ...)

\method{data_plot}{p_direction}(x, data = NULL, ...)

\method{plot}{p_direction}(x, data = NULL, ...)

\method{data_plot}{rope}(x, data = NULL, ...)

\method{plot}{rope}(x, data = NULL, rope_alpha = 0.5,
  rope_color = "grey", ...)
}
\arguments{
\item{x}{An object.}

\item{data}{The original data used to create this object. Can be a statistical model or such.}

\item{...}{Arguments passed to or from other methods.}

\item{rope_alpha}{Transparency level of ROPE ribbon.}

\item{rope_color}{Color of ROPE ribbon.}
}
\description{
This function attempts to extract and tranform an object to be further plotted.
}
\examples{
\dontrun{
library(rstanarm)
library(estimate)

model <- stan_glm(Sepal.Width ~ Species, data=iris)

contrasts <- estimate_contrasts(model, test=NULL)
means <- estimate_means(model)

data <- data_plot(contrasts, means)

plot(data)
}
\dontrun{
library(rstanarm)
library(estimate)

model <- stan_glm(Sepal.Width ~ Species, data=iris)

contrasts <- estimate_contrasts(model, test=NULL)
means <- estimate_means(model)

plot(contrasts, means)
}

library(bayestestR)
library(see)

data <- rnorm(1000, 1)
x <- hdi(data, c(0.8, 0.9))
data <- data_plot(x, data)
plot(data)

\dontrun{
library(rstanarm)
model <- rstanarm::stan_glm(Sepal.Length ~ Petal.Width * Species, data=iris)
x <- hdi(model)
data <- data_plot(x)
plot(data)
}
\dontrun{
library(bayestestR)
data <- rnorm(1000, 1)
x <- hdi(data, ci=c(0.8, 0.9))

plot(x, data) +
  theme_modern()

}

library(bayestestR)
library(see)

data <- rnorm(1000, 1)
x <- p_direction(data)
data <- data_plot(x, data)
plot(data)

\dontrun{
library(rstanarm)
model <- rstanarm::stan_glm(Sepal.Length ~ Petal.Width * Species, data=iris)
x <- p_direction(model)
data <- data_plot(x)
plot(data)
}
library(bayestestR)
data <- rnorm(1000, 1)
x <- p_direction(data)

plot(x, data) +
  theme_modern()


library(bayestestR)
library(see)

data <- rnorm(1000, 1)

x <- rope(data)
dataplot <- data_plot(x, data)
plot(dataplot)

x <- rope(data, ci=c(0.8, 0.9))
dataplot <- data_plot(x, data)
plot(dataplot)

\dontrun{
library(rstanarm)
data <- rstanarm::stan_glm(Sepal.Length ~ Petal.Width * Species, data=iris)
x <- rope(data)
dataplot <- data_plot(x, data)
plot(dataplot)

x <- rope(data, ci=c(0.8, 0.9))
dataplot <- data_plot(x, data)
plot(dataplot)
}
\dontrun{
# library(bayestestR)
# data <- rnorm(1000, 1)
# x <- rope(data, ci=c(0.8, 0.9))

# plot(x, data) +
#   theme_modern()

# data <- rstanarm::stan_glm(Sepal.Length ~ Petal.Width * Species, data=iris)
# x <- rope(data, ci=c(0.8, 0.9))
}
}
