% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.effectsize_table.R
\docType{data}
\name{.es_names}
\alias{.es_names}
\title{List of effect size names}
\format{
An object of class \code{list} of length 2.
}
\usage{
.es_names
}
\description{
List of effect size names
}
\keyword{internal}
