% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_flat.R
\name{flat_colors}
\alias{flat_colors}
\title{Extract Flat UI colors as hex codes}
\usage{
flat_colors(...)
}
\arguments{
\item{...}{Character names of colors.}
}
\value{
A character vector with color-codes.
}
\description{
Can be used to get the hex code of specific colors from the Flat UI color
palette. Use \code{flat_colors()} to see all available colors.
}
\examples{
flat_colors()

flat_colors("dark red", "teal")
}
