% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golden_ratio.R
\name{golden_ratio}
\alias{golden_ratio}
\title{Golden Ratio}
\usage{
golden_ratio(x = 1)
}
\arguments{
\item{x}{A number to be multiplied by the golden ratio. The default (\code{x = 1})
returns the value of the golden ratio.}
}
\description{
Returns the golden ratio (1.618034...). Useful to easily obtain golden
proportions, for instance for a horizontal figure, if you want its height to
be 8, you can set its width to be \code{golden_ratio(8)}.
}
\examples{
golden_ratio()
golden_ratio(10)
}
