% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_material.R
\name{material_colors}
\alias{material_colors}
\title{Extract material design colors as hex codes}
\usage{
material_colors(...)
}
\arguments{
\item{...}{Character names of \href{https://www.materialui.co/colors}{material design colors}.}
}
\description{
Extract material design colors as hex codes
}
