% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_metro.R
\name{metro_colors}
\alias{metro_colors}
\title{Extract Metro colors as hex codes}
\usage{
metro_colors(...)
}
\arguments{
\item{...}{Character names of colors.}
}
\value{
A character vector with color-codes.
}
\description{
Can be used to get the hex code of specific colors from the Metro color
palette. Use \code{metro_colors()} to see all available colors.
}
\examples{
metro_colors()

metro_colors("dark red", "teal")
}
