% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_material.R
\name{palette_material}
\alias{palette_material}
\title{Material design color palette}
\usage{
palette_material(palette = "contrast", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in material_palettes. Can be "full", "ice", "rainbow" or "contrast" (default).}

\item{reverse}{Boolean indicating whether the palette should be reversed.}

\item{...}{Additional arguments to pass to \code{\link[=colorRampPalette]{colorRampPalette()}}.}
}
\description{
The palette based on material design colors (https://www.materialui.co/colors).
}
