% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_metro.R
\name{palette_metro}
\alias{palette_metro}
\title{Metro color palette}
\usage{
palette_metro(palette = "complement", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette. Can be "full", "ice", "rainbow", "complement", or "contrast".}

\item{reverse}{Boolean indicating whether the palette should be reversed.}

\item{...}{Additional arguments to pass to \code{\link[=colorRampPalette]{colorRampPalette()}}.}
}
\description{
The palette based on Metro colors (https://www.materialui.co/metrocolors).
}
\details{
This function is usually not called directly, but from within
  \code{\link[=scale_color_metro]{scale_color_metro()}}.
}
