% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_see.R
\name{palette_see}
\alias{palette_see}
\title{see design color palette}
\usage{
palette_see(palette = "contrast", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette. Can be "full", "ice", "rainbow", "complement", or "contrast".}

\item{reverse}{Boolean indicating whether the palette should be reversed.}

\item{...}{Additional arguments to pass to \code{\link[=colorRampPalette]{colorRampPalette()}}.}
}
\description{
see design color palette.
}
\details{
This function is usually not called directly, but from within
  \code{\link[=scale_color_see]{scale_color_see()}}.
}
