% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_tol.R
\name{palette_tol_discrete}
\alias{palette_tol_discrete}
\title{Paul Tol's color palettes}
\usage{
palette_tol_discrete(palette = "bright", reverse = FALSE, order = NULL, ...)
}
\arguments{
\item{palette}{Character name of palette. Can be:
\itemize{
\item Qualitative: \code{"bright"}, \code{"high-contrast"}, \code{"vibrant"}, \code{"muted"},
\code{"medium-contrast"}, \code{"pale"}, \code{"dark"}, \code{"light"}, \code{"ground_cover"}
\item Diverging: \code{"sunset"}, \code{"BuRd"}, \code{"PRGn"}
\item Sequential: \code{"YlOrBr"}, \code{"iridescent"}, \code{"rainbow_discrete"},
\item \code{"rainbow_smooth"}
}}

\item{reverse}{Boolean indicating whether the palette should be reversed.}

\item{order}{A vector of numbers indicating the order of colors to use
(default: \code{NULL} indicating to use all available colors in order).}

\item{...}{For sequential palettes other than \code{rainbow_discrete}, additional
arguments to pass to \code{\link[grDevices:colorRamp]{grDevices::colorRampPalette()}}.}
}
\description{
The palettes proposed by Tol (2021).
}
\details{
This function is usually not called directly, but from within
\code{\link[=scale_color_tol]{scale_color_tol()}}.
}
\references{
Tol, P. (2021). Colour schemes (SRON/EPS Technical Note No. 09-002; Version 3.2).
SRON. https://personal.sron.nl/~pault/data/colourschemes.pdf (Original work published 2009)
}
