% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_pizza.R
\name{pizza_colors}
\alias{pizza_colors}
\title{Extract pizza colors as hex codes}
\usage{
pizza_colors(...)
}
\arguments{
\item{...}{Character names of pizza ingredients.}
}
\description{
Extract pizza colors as hex codes
}
