% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.dw_data_tabulate.R
\name{plot.dw_data_tabulates}
\alias{plot.dw_data_tabulates}
\alias{plot.dw_data_tabulate}
\title{Plot tabulated data.}
\usage{
\method{plot}{dw_data_tabulates}(
  x,
  value_lab = TRUE,
  remove_na = FALSE,
  na_label = "(Missing)",
  error_bar = TRUE,
  ...
)

\method{plot}{dw_data_tabulate}(
  x,
  value_lab = TRUE,
  remove_na = FALSE,
  na_label = "(Missing)",
  error_bar = TRUE,
  ...
)
}
\arguments{
\item{x}{Object created by \code{datawizard::data_tabulate()}.}

\item{value_lab}{Boolean. Should values and percentages be displayed at the
top of each bar.}

\item{remove_na}{Boolean. Should missing values be dropped?}

\item{na_label}{The label given to missing values (only when
\code{remove_na = FALSE}).}

\item{error_bar}{Boolean. Should error bars be displayed?}

\item{...}{Unused}
}
\description{
Plot tabulated data.
}
