% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.dw_groupmeans.R
\name{plot.dw_groupmeans}
\alias{plot.dw_groupmeans}
\title{Plot group means}
\usage{
\method{plot}{dw_groupmeans}(x, y, ...)
}
\arguments{
\item{x}{A dw_groupmeans data frame.}

\item{y}{Always \code{NULL} for this plot type}

\item{...}{Arguments to be passed to methods. \code{title}
can be used to specify a string to use a title. \code{caption}
is a logical indicating if a caption summarizing the anova
results for the analysis should be included. By default this
is \code{TRUE}. \code{ci} is a logical indicating if the confidence
intervals from the \code{x} dw_groupmeans object should be included in the graph.}
}
\description{
Plot group means
}
\examples{
group_means_object <-  datawizard::means_by_group(iris$Sepal.Width, iris$Species)
plot(group_means_object, title = "group means", ci = FALSE, caption = FALSE)

}
