% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.dw_groupmeans.R
\name{plot.dw_groupmeans_list}
\alias{plot.dw_groupmeans_list}
\title{function to plot lists of \code{dw_groupmeans}}
\usage{
\method{plot}{dw_groupmeans_list}(x, y, ...)
}
\arguments{
\item{x}{A dw_groupmeans data frame.}

\item{...}{Other options. The function recognizes \code{title},
\code{caption}, and \code{ci }. \code{title} is an overall title.}
}
\description{
function to plot lists of \code{dw_groupmeans}
}
\details{
Produces a faceted plot when there is more than one
\code{dwtable}
}
\examples{
group_means_object <-  datawizard::means_by_group(iris, Sepal.Width, Species)
plot(group_means_object, title = "group means", ci = FALSE, caption = FALSE)


}
