% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.check_overdisp.R
\name{plot.see_check_overdisp}
\alias{plot.see_check_overdisp}
\title{Plot method to check model for overdispersion}
\usage{
\method{plot}{see_check_overdisp}(x, size_line = 0.8, colors = c("#3aaf85", "#1b6ca8"), ...)
}
\arguments{
\item{x}{An object.}

\item{size_line}{Numeric value specifying size of line geoms.}

\item{colors}{Character vector of length two, indicating the colors (in
hex-format) for points and line.}

\item{...}{Arguments passed to or from other methods.}

\item{type}{Character vector, indicating the type of plot.}

\item{dot_alpha, alpha}{Numeric value specifying alpha level of the
confidence bands and point-geoms.}

\item{detrend}{Logical that decides if the plot should be detrended.}
}
\value{
A ggplot2-object.
}
\description{
The \code{plot()} method for the \code{performance::check_overdispersion()}
function.
}
\examples{
library(performance)
m <<- lm(mpg ~ wt + cyl + gear + disp, data = mtcars)
result <- check_normality(m)
plot(result)
}
