% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.check_priors.R
\name{plot.see_check_priors}
\alias{plot.see_check_priors}
\title{Plot method for prior predictive checks}
\usage{
\method{plot}{see_check_priors}(
  x,
  size_point = 2,
  size_boxplot = 0.4,
  size_title = 12,
  size_axis_title = base_size,
  base_size = 10,
  alpha_dot = 0.15,
  alpha_boxplot = 0.35,
  theme = NULL,
  ...
)
}
\arguments{
\item{x}{An object.}

\item{size_point}{Numeric specifying size of point-geoms.}

\item{size_boxplot}{Numeric value specifying size of boxplot geoms.}

\item{base_size, size_axis_title, size_title}{Numeric value specifying size of
axis and plot titles.}

\item{alpha_dot}{Numeric value specifying alpha level of the point geoms.}

\item{alpha_boxplot}{Numeric value specifying alpha of boxplot geoms.}

\item{theme}{A ggplot2-theme function, e.g. \code{theme = theme_lucid()} or
\code{theme = ggplot2::theme_dark()}.}

\item{...}{Arguments passed to or from other methods.}
}
\value{
A ggplot2-object.
}
\description{
The \code{plot()} method for the \code{performance::check_priors()} function. For
details, see \emph{Lüdecke et al. 2026}.
}
\examples{
\dontrun{
library(performance)
# model with correctly defined priors. outcome is binary, prior
# predictive checks indicate the predicted probability mass based
# on the prior distributions - the resulting pattern aligns with
# our real-world assumptions
model <- insight::download_model("stan_prior_checks_1")
plot(performance::check_priors(model, "mmse"))

# model with default (weakly informative) priors, which is poorly
# calibrated. It pushes probability mass almost exclusively to the
# extremes of 0 and 1, leaving the plausible middle range largely
# unsupported
model <- insight::download_model("stan_prior_checks_2")
plot(performance::check_priors(model, "mmse"))
}

}
\references{
Lüdecke D, Makowski AC, Klein J, Ben-Shachar MS and Makowski D (2026)
Choosing informative priors in Bayesian regression models: a simulation study
and tutorial using Stan and R. Front. Psychol. 17:1856582.
\doi{10.3389/fpsyg.2026.1856582}
}
