% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.compare_performance.R
\name{plot.see_compare_performance}
\alias{plot.see_compare_performance}
\title{Plot method for comparing model performances}
\usage{
\method{plot}{see_compare_performance}(x, size_line = 1, ...)
}
\arguments{
\item{x}{An object.}

\item{size_line}{Numeric value specifying size of line geoms. \code{linewidth} is
an alias for \code{size_line}.}

\item{...}{Arguments passed to or from other methods.}
}
\value{
A ggplot2-object.
}
\description{
The \code{plot()} method for the \code{performance::compare_performance()} function.
}
\examples{
library(performance)
data(iris)
lm1 <- lm(Sepal.Length ~ Species, data = iris)
lm2 <- lm(Sepal.Length ~ Species + Petal.Length, data = iris)
lm3 <- lm(Sepal.Length ~ Species * Petal.Length, data = iris)
result <- compare_performance(lm1, lm2, lm3)
result
plot(result)

}
