% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.easycormatrix.R
\name{plot.see_easycormatrix}
\alias{plot.see_easycormatrix}
\title{Plot method for correlation matrices}
\usage{
\method{plot}{see_easycormatrix}(
  x,
  show_values = FALSE,
  show_p = FALSE,
  show_legend = TRUE,
  size_point = 1,
  size_text = 3.5,
  digits = 3,
  type = c("circle", "tile"),
  ...
)
}
\arguments{
\item{x}{An object.}

\item{show_values}{Logical. If \code{TRUE}, values are displayed.}

\item{show_p}{Logical. If \code{TRUE}, \emph{p}-values or significant level is
displayed.}

\item{show_legend}{Logical, show (\code{TRUE}) or hide (\code{FALSE}) legend.}

\item{size_point}{Numeric specifying size of point-geoms.}

\item{size_text}{Numeric value specifying size of text labels.}

\item{digits}{Number of decimals used for values.}

\item{type}{Character vector, indicating the type of plot.}

\item{...}{Arguments passed to or from other methods.}
}
\value{
A ggplot2-object.
}
\description{
The \code{plot()} method for the \code{correlation::correlation()} function.
}
\examples{
library(correlation)
data(mtcars)
result <- correlation(mtcars[, -c(8:9)])
s <- summary(result)
plot(s)
}
