% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.parameters_pca.R
\name{plot.see_parameters_pca}
\alias{plot.see_parameters_pca}
\title{Plot method for principal component analysis}
\usage{
\method{plot}{see_parameters_pca}(
  x,
  type = "bar",
  size_text = 3.5,
  color_text = "black",
  colors = c("#cd201f", "#ffffff", "#0077B5"),
  size = 1,
  show_labels = TRUE,
  ...
)
}
\arguments{
\item{x}{An object.}

\item{type}{Character vector, indicating the type of plot. Options are
different shapes to represent component loadings; \code{"bar"} (default) for a
horizontal bar chart, \code{"line"} for a horizontal point and line chart, or
\code{"graph"} for a graph.}

\item{size_text}{Numeric value specifying size of text labels.}

\item{color_text}{Character specifying color of text labels.}

\item{colors}{Character vector of length three, indicating the colors (in
hex-format) for low (negative), mid (close to zero), and high (positive) values.}

\item{size}{Depending on \code{type}, a numeric value specifying size of bars,
lines, or segments.}

\item{show_labels}{Logical. If \code{TRUE}, text labels are displayed.}

\item{...}{Arguments passed to or from other methods.}
}
\value{
A ggplot2-object.
}
\description{
The \code{plot()} method for the \code{parameters::principal_components()} function.
}
\examples{
library(parameters)
data(mtcars)
result <- principal_components(mtcars[, 1:7], n = "all", threshold = 0.2)
result
plot(result)

}
