% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_history.R
\name{scale_color_tate}
\alias{scale_color_tate}
\alias{scale_color_tate_d}
\alias{scale_color_tate_c}
\alias{scale_colour_tate}
\alias{scale_colour_tate_c}
\alias{scale_colour_tate_d}
\alias{scale_fill_tate}
\alias{scale_fill_tate_d}
\alias{scale_fill_tate_c}
\title{Tate Modern color palette}
\usage{
scale_color_tate(
  palette = "history",
  discrete = TRUE,
  reverse = FALSE,
  aesthetics = "color",
  ...
)

scale_color_tate_d(
  palette = "history",
  discrete = TRUE,
  reverse = FALSE,
  aesthetics = "color",
  ...
)

scale_color_tate_c(
  palette = "history",
  discrete = FALSE,
  reverse = FALSE,
  aesthetics = "color",
  ...
)

scale_colour_tate(
  palette = "history",
  discrete = TRUE,
  reverse = FALSE,
  aesthetics = "color",
  ...
)

scale_colour_tate_c(
  palette = "history",
  discrete = FALSE,
  reverse = FALSE,
  aesthetics = "color",
  ...
)

scale_colour_tate_d(
  palette = "history",
  discrete = TRUE,
  reverse = FALSE,
  aesthetics = "color",
  ...
)

scale_fill_tate(
  palette = "history",
  discrete = TRUE,
  reverse = FALSE,
  aesthetics = "fill",
  ...
)

scale_fill_tate_d(
  palette = "history",
  discrete = TRUE,
  reverse = FALSE,
  aesthetics = "fill",
  ...
)

scale_fill_tate_c(
  palette = "history",
  discrete = FALSE,
  reverse = FALSE,
  aesthetics = "fill",
  ...
)
}
\arguments{
\item{palette}{Character name of palette. Depending on the color scale, can
be \code{"full"}, \code{"ice"}, \code{"rainbow"}, \code{"complement"},
\code{"contrast"}, \code{"light"} (for dark themes), \code{"black_first"}, \code{full_original},
or \code{black_first_original}.}

\item{discrete}{Boolean indicating whether color aesthetic is discrete or not.}

\item{reverse}{Boolean indicating whether the palette should be reversed.}

\item{aesthetics}{A vector of names of the aesthetics that this scale
should be applied to (e.g., \code{c('color', 'fill')}).}

\item{...}{Additional arguments to pass to \code{\link[=colorRampPalette]{colorRampPalette()}}.}
}
\description{
A color palette inspired by paintings in the Tate Modern Art Museum in London.
Use \code{scale_color_tate_d()} for \emph{discrete} categories and
\code{scale_color_tate_c()} for a \emph{continuous} scale.
}
\examples{
library(ggplot2)
library(see)

ggplot(iris, aes(x = Species, y = Sepal.Length, fill = Species)) +
  geom_boxplot() +
  theme_modern() +
  scale_fill_tate_d()

ggplot(iris, aes(x = Sepal.Length, y = Sepal.Width, colour = Species)) +
  geom_point() +
  theme_abyss() +
  scale_colour_tate(palette = "fontana")
}
