% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/see-package.R
\docType{package}
\name{see-package}
\alias{see-package}
\alias{see}
\title{see: Framework for Easy Statistical Modeling, Visualization, and Reporting}
\description{
Provides plotting utilities supporting packages in the 'easystats' ecosystem
(\url{https://github.com/easystats/easystats}) and some extra themes, geoms, and
scales for 'ggplot2'. Color scales are based on
\url{https://materialui.co/}.

References: Luedecke et al. (2021) \doi{10.21105/joss.03393}.
}
\details{
\code{see}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://easystats.github.io/see/}
  \item Report bugs at \url{https://github.com/easystats/see/issues}
}

}
\author{
\strong{Maintainer}: Daniel Lüdecke \email{officialeasystats@gmail.com} (\href{https://orcid.org/0000-0002-8895-3206}{ORCID})

Authors:
\itemize{
  \item Daniel Lüdecke \email{officialeasystats@gmail.com} (\href{https://orcid.org/0000-0002-8895-3206}{ORCID})
  \item Dominique Makowski \email{dom.makowski@gmail.com} (\href{https://orcid.org/0000-0001-5375-9967}{ORCID}) [inventor]
  \item Indrajeet Patil \email{patilindrajeet.science@gmail.com} (\href{https://orcid.org/0000-0003-1995-6531}{ORCID})
  \item Mattan S. Ben-Shachar \email{matanshm@post.bgu.ac.il} (\href{https://orcid.org/0000-0002-4287-4801}{ORCID}) [contributor]
  \item Brenton M. Wiernik \email{brenton@wiernik.org} (\href{https://orcid.org/0000-0001-9560-6336}{ORCID}) [contributor]
  \item Rémi Thériault \email{remi.theriault@mail.mcgill.ca} (\href{https://orcid.org/0000-0003-4315-6788}{ORCID}) [contributor]
  \item Philip Waggoner \email{philip.waggoner@gmail.com} (\href{https://orcid.org/0000-0002-7825-7573}{ORCID}) [contributor]
}

Other contributors:
\itemize{
  \item Jeffrey R. Stevens \email{jeffrey.r.stevens@gmail.com} (\href{https://orcid.org/0000-0003-2375-1360}{ORCID}) [contributor]
  \item Julius Bogomolovas (\href{https://orcid.org/0000-0001-8344-1909}{ORCID}) [contributor]
  \item Matthew Smith \email{M.Smith3@napier.ac.uk} [reviewer]
  \item Jakob Bossek \email{bossek@wi.uni-muenster.de} [reviewer]
}

}
\keyword{internal}
