% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_see.R
\name{see_colors}
\alias{see_colors}
\title{Extract See colors as hex codes}
\usage{
see_colors(...)
}
\arguments{
\item{...}{Character names of colors.}
}
\value{
A character vector with color-codes.
}
\description{
Can be used to get the hex code of specific colors from the See color palette. Use \code{see_colors()} to see all available color.
}
\examples{
see_colors()

see_colors("indigo", "lime")
}
