% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_social.R
\name{social_colors}
\alias{social_colors}
\title{Extract Social colors as hex codes}
\usage{
social_colors(...)
}
\arguments{
\item{...}{Character names of colors.}
}
\value{
A character vector with color-codes.
}
\description{
Can be used to get the hex code of specific colors from the Social color
palette. Use \code{social_colors()} to see all available colors.
}
\examples{
social_colors()

social_colors("dark red", "teal")
}
