% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_history.R
\name{tate_colors}
\alias{tate_colors}
\title{Extract tate colors as hex codes}
\usage{
tate_colors(...)
}
\arguments{
\item{...}{Character names of colors.}
}
\value{
A character vector with color-codes.
}
\description{
Can be used to get the hex code of specific colors from the tate color
palette. Use \code{tate_colors()} to see all available colors.
}
\examples{
tate_colors()

tate_colors("indigo", "lime")
}
