% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_light.R
\name{theme_light}
\alias{theme_light}
\title{Light theme}
\usage{
theme_light(base_size = 11, base_family = "")
}
\description{
A light theme for ggplot.
}
\examples{
library(ggplot2)
library(see)

ggplot(iris, aes(x = Sepal.Width, y = Sepal.Length)) +
  geom_point(color = "white") +
  theme_light()
}
