% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_modern.R
\name{theme_modern}
\alias{theme_modern}
\title{The easystats' minimal theme}
\usage{
theme_modern(legend.position = "right", axis.title.space = 20,
  legend.title.size = 13, legend.text.size = 12,
  axis.title.size = 13, axis.title.face = "plain",
  axis.text.size = 12, tags.size = 15, tags.face = "bold")
}
\arguments{
\item{legend.position}{the position of legends ("none", "left", "right",
"bottom", "top", or two-element numeric vector)}

\item{axis.title.space}{Axis title spacing.}

\item{legend.title.size}{Legend elements text size in pts.}

\item{legend.text.size}{Legend elements text size in pts. Can be "none".}

\item{axis.title.size}{Axis title text size in pts.}

\item{axis.title.face}{Axis font face ("plain", "italic", "bold", "bold.italic").}

\item{axis.text.size}{Axis text size in pts.}

\item{tags.size}{Tags text size in pts.}

\item{tags.face}{Tags font text ("plain", "italic", "bold", "bold.italic").}
}
\description{
A modern, sleek and elegant theme for ggplot.
}
\examples{
library(ggplot2)
library(see)

ggplot(iris, aes(x = Sepal.Width, y = Sepal.Length, color = Species)) +
  geom_point() +
  theme_modern()
}
