% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_pizza.R
\docType{data}
\name{theme_pizza}
\alias{theme_pizza}
\title{Pizza theme}
\format{An object of class \code{list} of length 3.}
\usage{
theme_pizza
}
\description{
A theme based on authentic neapolitan pizzas.
}
\examples{
library(ggplot2)
library(see)

ggplot(iris, aes(x = Sepal.Width, y = Sepal.Length)) +
  geom_point() +
  theme_pizza()
}
\keyword{datasets}
