% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_tol.R
\name{tol_colors}
\alias{tol_colors}
\title{Extract Paul Tol colors as hex codes}
\usage{
tol_colors(..., palette = "bright")
}
\arguments{
\item{...}{Character names of colors.}

\item{palette}{Character name of palette. Can be:
\itemize{
\item Qualitative: \code{"bright"}, \code{"high-contrast"}, \code{"vibrant"}, \code{"muted"},\code{ "medium-contrast"}, \code{"pale"}, \code{"dark"}, \code{"light"}, \code{"ground_cover"}
\item Diverging: \code{"sunset"}, \code{"BuRd"}, \code{"PRGn"}
\item Sequential: \code{"YlOrBr"}, \code{"iridescent"}, \code{"rainbow_discrete"},
\code{"rainbow_smooth"}
}}
}
\value{
A character vector with color-codes.
}
\description{
Can be used to get the hex code of specific colors from the Paul Tol palettes.
Use \code{tol_colors()} and specify \code{palette} to see all available colors.
Note that for sequential palettes, only original (non-interpolated) colors are shown.
}
\examples{
tol_colors()

tol_colors(c("red", "light blue", "yellow"))

tol_colors(palette = "muted")

tol_colors(c("red", "light blue", "yellow"), palette = "muted")
}
