% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_uke.R
\name{uke_colors}
\alias{uke_colors}
\title{Extract UKE colors as hex codes}
\usage{
uke_colors(...)
}
\arguments{
\item{...}{Character names of colors.}
}
\value{
A character vector with color-codes.
}
\description{
Can be used to get the hex code of specific colors from the UKE color
palette. Use \code{uke_colors()} to see all available colors.
}
\examples{
uke_colors()

uke_colors("blue", "cyan")
}
