% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_plot.R, R/plot.rope.R
\name{data_plot}
\alias{data_plot}
\alias{plot.rope}
\title{Prepare an object for plotting}
\usage{
data_plot(x, data = NULL, ...)

\method{plot}{rope}(x, data = NULL, rope_alpha = 0.5,
  rope_color = "cadetblue", ...)
}
\arguments{
\item{x}{An object.}

\item{data}{The original data used to create this object. Can be a statistical model or such.}

\item{...}{Arguments passed to or from other methods.}

\item{rope_alpha}{Transparency level of ROPE ribbon.}

\item{rope_color}{Color of ROPE ribbon.}
}
\description{
This function attempts to extract and tranform an object to be further plotted.
}
\examples{
library(bayestestR)

data <- rnorm(1000, 1)
x <- rope(data, ci = c(0.8, 0.9))
plot(x)

\dontrun{
library(rstanarm)
model <- stan_glm(Sepal.Length ~ Petal.Width * Species, data = iris)
x <- rope(model)
plot(x)
}

data <- rnorm(1000, 1)
x <- hdi(data, c(0.8, 0.9))
plot(x) + theme_modern()

\dontrun{
library(rstanarm)
model <- stan_glm(Sepal.Length ~ Petal.Width * Species, data = iris)
x <- hdi(model)
plot(x) + theme_modern()
}

data <- rnorm(1000, 1)
x <- p_direction(data)
plot(x)

\dontrun{
library(rstanarm)
model <- stan_glm(Sepal.Length ~ Petal.Width * Species, data = iris)
x <- p_direction(model)
plot(x)
}

x <- equivalence_test(iris, ci = c(.8, .9))
plot(x) +
  theme_blackboard() +
  scale_fill_material()

\dontrun{
library(rstanarm)
model <- stan_glm(Sepal.Length ~ Petal.Width * Species, data = iris)
x <- equivalence_test(model)
plot(x)
}

}
