% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayestestR_plot.p_direction.R, R/data_plot.R
\name{data_plot.p_direction}
\alias{data_plot.p_direction}
\alias{plot.p_direction}
\alias{data_plot}
\title{Prepare an object for plotting}
\usage{
\method{data_plot}{p_direction}(x, data = NULL, ...)

\method{plot}{p_direction}(x, data = NULL, ...)

data_plot(x, data = NULL, ...)
}
\arguments{
\item{x}{An object.}

\item{data}{The original data used to create this object. Can be a statistical model or such.}

\item{...}{Arguments passed to or from other methods.}
}
\description{
This function attempts to extract and tranform an object to be further plotted.
}
\examples{
\dontrun{
library(bayestestR)
library(see)

data <- rnorm(1000, 1)
x <- p_direction(data)
data <- data_plot(x, data)
plot(data)

library(rstanarm)
model <- rstanarm::stan_glm(Sepal.Length ~ Petal.Width * Species, data=iris)
x <- p_direction(model)
data <- data_plot(x)
plot(data)
}
\dontrun{
library(bayestestR)
data <- rnorm(1000, 1)
x <- p_direction(data)

plot(x, data) +
  theme_modern()

}
}
