% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_from_list.R
\name{geom_from_list}
\alias{geom_from_list}
\title{Create ggplot2 geom(s) from a list}
\usage{
geom_from_list(x, ...)
}
\arguments{
\item{x}{A list containing a geom type (e.g., \code{geom = "point"}), a list of aesthetics (as characters; e.g., \code{aes = list(x = "mpg", y = "wt")}), some data (e.g., \code{data = mtcars}) and some other parameters. For \code{geoms_from_list()} ("geoms" with an "s"), the input must be a list of lists, ideally named \code{"l1", "l2", "l3"}, etc.}
}
\description{
These helper functions are built on top of \code{ggplot2::layer()} and can be used
to add geom(s) which type and content is specified as a list.
}
\examples{
library(ggplot2)

l1 <- list(geom = "point",
           data = mtcars,
           aes = list(x = "mpg", y = "wt"))
l2 <- list(geom = "labs",
           title = "A Title")

ggplot() +
  geom_from_list(l1) +
  geom_from_list(l2)

ggplot() +
  geoms_from_list(list(l1 = l1, l2 = l2))


}
