% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/how_to_plot.R
\name{how_to_plot}
\alias{how_to_plot}
\title{How to plot your object}
\usage{
how_to_plot(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Arguments passed to or from other methods.}
}
\description{
Access the source code used for plotting.
}
\examples{
library(bayestestR)
results <- hdi(rnorm(1000))
how_to_plot(results)
}
