% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_material.R
\name{material_colors}
\alias{material_colors}
\title{Extract material design colors as hex codes}
\usage{
material_colors(...)
}
\arguments{
\item{...}{Character names of colors.}
}
\value{
A character vector with color-codes.
}
\description{
Can be used to get the hex code of specific colors from the material design
color palette. Use \code{material_colors()} to see all available colors.
}
\examples{
material_colors()

material_colors("indigo", "lime")
}
