% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_metro.R
\name{palette_metro}
\alias{palette_metro}
\title{Metro color palette}
\usage{
palette_metro(palette = "complement", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette. Depending on the color scale, can
be \code{"full"}, \code{"ice"}, \code{"rainbow"}, \code{"complement"},
\code{"contrast"}, \code{"light"} (for dark themes), \code{"black_first"}, \code{full_original},
or \code{black_first_original}.}

\item{reverse}{Boolean indicating whether the palette should be reversed.}

\item{...}{Additional arguments to pass to \code{\link[=colorRampPalette]{colorRampPalette()}}.}
}
\description{
The palette based on \href{https://materialui.co/metrocolors}{Metro colors}.
}
\details{
This function is usually not called directly, but from within
\code{\link[=scale_color_metro]{scale_color_metro()}}.
}
