% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_social.R
\name{palette_social}
\alias{palette_social}
\title{Social color palette}
\usage{
palette_social(palette = "complement", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette. Depending on the color scale, can be \code{"full"}, \code{"ice"}, \code{"rainbow"}, \code{"complement"}, \code{"contrast"} or \code{"light"} (for dark themes).}

\item{reverse}{Boolean indicating whether the palette should be reversed.}

\item{...}{Additional arguments to pass to \code{\link[=colorRampPalette]{colorRampPalette()}}.}
}
\description{
The palette based on Social colors (https://www.materialui.co/socialcolors).
}
\details{
This function is usually not called directly, but from within
  \code{\link[=scale_color_social]{scale_color_social()}}.
}
