% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.dw_groupmeans.R
\name{plot.dw_groupmeans}
\alias{plot.dw_groupmeans}
\title{Plot method for grouped means}
\usage{
\method{plot}{dw_groupmeans}(x, ...)
}
\arguments{
\item{x}{An object returned \code{datawizard::means_by_group.data()}.}

\item{...}{Additional options. Acceptable values include \code{title}
with a string to use as a title, \code{ci} for if confidence intervals
(if present in the object should be graphed), and \code{caption} for if
a caption summarizing ANOVA results should be included.}

\item{y}{Not currently used}
}
\description{
The \code{plot()} method for the \code{datawizard::means_by_group()} function
}
\examples{
group_means_object <-  datawizard::means_by_group(iris$Sepal.Width, iris$Species)
plot(group_means_object, title = "group means", ci = FALSE, caption = FALSE)

group_means_object <- datawizard::means_by_group(
  iris,
  c("Sepal.Width", "Petal.Width"),
  "Species"
)

plot(group_means_object, title = "group means")

group_means_object <- datawizard::means_by_group(
  iris$Sepal.Width, iris$Species)
)
plot(group_means_object, title = "group means")

}
