% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.check_collinearity.R
\name{plot.see_check_collinearity}
\alias{plot.see_check_collinearity}
\title{Plot method for multicollinearity checks}
\usage{
\method{plot}{see_check_collinearity}(
  x,
  data = NULL,
  theme = NULL,
  colors = c("#3aaf85", "#1b6ca8", "#cd201f"),
  size_point = 3.5,
  size_line = 0.8,
  size_title = 12,
  size_axis_title = base_size,
  base_size = 10,
  ...
)
}
\arguments{
\item{x}{An object.}

\item{data}{The original data used to create this object. Can be a
statistical model.}

\item{theme}{A ggplot2-theme function, e.g. \code{theme = theme_lucid()} or
\code{theme = ggplot2::theme_dark()}.}

\item{colors}{Character vector of length two, indicating the colors (in
hex-format) for points and line.}

\item{size_point}{Numeric specifying size of point-geoms.}

\item{size_line}{Numeric value specifying size of line geoms. \code{linewidth} is
an alias for \code{size_line}.}

\item{base_size, size_axis_title, size_title}{Numeric value specifying size of
axis and plot titles.}

\item{...}{Arguments passed to or from other methods.}
}
\value{
A ggplot2-object.
}
\description{
The \code{plot()} method for the \code{performance::check_collinearity()} function.
}
\examples{
library(performance)
m <- lm(mpg ~ wt + cyl + gear + disp, data = mtcars)
result <- check_collinearity(m)
result
plot(result)

}
