% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.check_dag.R
\name{plot.see_check_dag}
\alias{plot.see_check_dag}
\title{Plot method for check DAGs}
\usage{
\method{plot}{see_check_dag}(
  x,
  size_point = 15,
  colors = NULL,
  which = "all",
  check_colliders = TRUE,
  ...
)
}
\arguments{
\item{x}{A \code{check_dag} object.}

\item{size_point}{Numeric value specifying size of point geoms.}

\item{colors}{Character vector of length five, indicating the colors (in
hex-format) for different types of variables.}

\item{which}{Character string indicating which plot to show. Can be either
\code{"all"}, \code{"current"} or \code{"required"}.}

\item{check_colliders}{Logical indicating whether to highlight colliders.
Set to \code{FALSE} if the algorithm to detect colliders is very slow.}

\item{...}{Not used.}
}
\value{
A ggplot2-object.
}
\description{
The \code{plot()} method for the \code{performance::check_dag()} function.
}
\examples{
\dontshow{if (require("ggdag", quietly = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
library(performance)
# incorrect adjustment
dag <- check_dag(
  y ~ x + b + c,
  x ~ b,
  outcome = "y",
  exposure = "x"
)
dag
plot(dag)

# plot only model with required adjustments
plot(dag, which = "required")

# collider-bias?
dag <- check_dag(
  y ~ x + c + d,
  x ~ c + d,
  b ~ x,
  b ~ y,
  outcome = "y",
  exposure = "x",
  adjusted = "c"
)
plot(dag)
\dontshow{\}) # examplesIf}
}
