% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.check_overdisp.R
\name{plot.see_check_overdisp}
\alias{plot.see_check_overdisp}
\title{Plot method to check model for overdispersion}
\usage{
\method{plot}{see_check_overdisp}(x, size_line = 0.8, colors = c("#3aaf85", "#1b6ca8"), ...)
}
\arguments{
\item{x}{An object.}

\item{size_line}{Numeric value specifying size of line geoms.}

\item{...}{Arguments passed to or from other methods.}
}
\value{
A ggplot2-object.
}
\description{
The \code{plot()} method for the \code{performance::check_overdispersion()}
function.
}
\examples{
library(performance)
if (require("glmmTMB")) {
  data(Salamanders)
  m <- glm(count ~ spp + mined, family = poisson, data = Salamanders)
  result <- check_overdispersion(m)
  plot(result)
}
}
