% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.means_by_group.R
\name{plot.see_dw_groupmeans}
\alias{plot.see_dw_groupmeans}
\title{Plot method for grouped means.}
\usage{
\method{plot}{see_dw_groupmeans}(x, title = "", ci = TRUE, caption = TRUE, ...)
}
\arguments{
\item{x}{An object returned \code{datawizard::means_by_group()}.}

\item{title}{String, can be used to specify a plot title.}

\item{ci}{Logical, indicating if the confidence intervals should be included
in the graph.}

\item{caption}{Logical, indicating if a caption summarizing the anova results
for the analysis should be included.}

\item{...}{Currently not used.}
}
\description{
The \code{plot()} method for the \code{datawizard::means_by_group()} function.
}
\details{
Produces a faceted plot when there is more than one means-table.
}
\examples{
\dontrun{
group_means_object <-  datawizard::means_by_group(iris$Sepal.Width, iris$Species)
plot(group_means_object, title = "group means", ci = FALSE, caption = FALSE)

group_means_object <- datawizard::means_by_group(
  iris,
  c("Sepal.Width", "Petal.Width"),
  "Species"
)
plot(group_means_object, title = "group means")
}
}
