% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.p_function.R
\name{plot.see_p_function}
\alias{plot.see_p_function}
\title{Plot method for plotting p-functions (aka consonance functions)}
\usage{
\method{plot}{see_p_function}(
  x,
  colors = c("black", "#1b6ca8"),
  size_points = 1.2,
  size_line = c(0.7, 0.9),
  size_text = 3,
  line_alpha = 0.15,
  show_values = TRUE,
  n_columns = NULL,
  show_intercept = FALSE,
  ...
)
}
\arguments{
\item{x}{An object returned by \code{parameters::p_function ()}.}

\item{colors}{Character vector of length two, indicating the colors (in
hex-format) used when only one parameter is plotted, resp. when panels
are plotted as facets.}

\item{show_intercept}{Logical, if \code{TRUE}, the intercept-parameter is included
in the plot. By default, it is hidden because in many cases the
intercept-parameter has a posterior distribution on a very different
location, so density curves of posterior distributions for other parameters
are hardly visible.}

\item{...}{Arguments passed to or from other methods.}
}
\value{
A ggplot2-object.
}
\description{
The \code{plot()} method for the \code{parameters::p_function ()}.
}
\examples{
library(performance)
m <<- lm(mpg ~ wt + cyl + gear + disp, data = mtcars)
result <- check_normality(m)
plot(result)
}
