% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.parameters_brms_meta.R
\name{plot.see_parameters_brms_meta}
\alias{plot.see_parameters_brms_meta}
\title{Plot method for Model Parameters from Bayesian Meta-Analysis}
\usage{
\method{plot}{see_parameters_brms_meta}(
  x,
  size_point = 1.5,
  size = 0.8,
  size_text = 3.5,
  rope_alpha = 0.15,
  rope_color = "cadetblue",
  ...
)
}
\arguments{
\item{x}{An object.}

\item{rope_alpha}{Transparency level of ROPE ribbon.}

\item{rope_color}{Color of ROPE ribbon.}

\item{...}{Arguments passed to or from other methods.}
}
\value{
A ggplot2-object.
}
\description{
The \code{plot()} method for the \code{parameters::model_parameters()} function when used with brms-meta-analysis models.
}
\details{
\subsection{Colors of density areas and errorbars}{
  To change the colors of the density areas, use \code{scale_fill_manual()}
  with named color-values, e.g. \code{scale_fill_manual(values = c("Study" = "blue", "Overall" = "green"))}.
  To change the color of the error bars, use \code{scale_color_manual(values = c("Errorbar" = "red"))}.
}
\subsection{Show or hide estimates and CI}{
  Use \code{size_text = NULL} or \code{size_text = NA} to hide the textual output of estimates and credible intervals.
}
}
\examples{
\dontrun{
if (require("bayestestR") && require("brms")) {
  # download data from:
  # https://github.com/MathiasHarrer/Doing-Meta-Analysis-in-R/blob/master/_data/Meta_Analysis_Data.RData
  set.seed(123)
  priors <- c(prior(normal(0,1), class = Intercept),
  prior(cauchy(0,0.5), class = sd))

  model <- brm(
    TE | se(seTE) ~ 1 + (1|Author),
    data = Meta_Analysis_Data,
    prior = priors,
    iter = 4000
  )

  mp <- model_parameters(model)
  plot(mp)
}
}
}
