% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.describe_distribution.R
\name{plot.see_parameters_distribution}
\alias{plot.see_parameters_distribution}
\title{Plot method for describing distributions of vectors}
\usage{
\method{plot}{see_parameters_distribution}(
  x,
  dispersion = FALSE,
  alpha_dispersion = 0.3,
  color_dispersion = "#3498db",
  dispersion_style = c("ribbon", "curve"),
  size_bar = 0.7,
  highlight = NULL,
  color_highlight = NULL,
  ...
)
}
\arguments{
\item{x}{An object.}

\item{dispersion}{Logical. If \code{TRUE}, a range of dispersion for
each variable to the plot will be added.}

\item{alpha_dispersion}{Numeric value specifying the transparency level of dispersion ribbon.}

\item{color_dispersion}{Character specifying the color of dispersion ribbon.}

\item{dispersion_style}{Character describing the style of dispersion area.
\code{"ribbon"} for a ribbon, \code{"curve"} for a normal-curve.}

\item{size_bar}{Size of bar geoms.}

\item{highlight}{A vector with names of categories in \code{x} that should be
highlighted.}

\item{color_highlight}{A vector of color values for highlighted categories.
The remaining (non-highlighted) categories will be filled with a lighter
grey.}

\item{...}{Arguments passed to or from other methods.}
}
\value{
A ggplot2-object.
}
\description{
The \code{plot()} method for the \code{parameters::describe_distribution()}
function.
}
\examples{
library(parameters)
set.seed(333)
x <- sample(1:100, 1000, replace = TRUE)
result <- describe_distribution(x)
result
plot(result)
}
