% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.check_predictions.R
\name{print.see_performance_pp_check}
\alias{print.see_performance_pp_check}
\alias{plot.see_performance_pp_check}
\title{Plot method for posterior predictive checks}
\usage{
\method{print}{see_performance_pp_check}(
  x,
  size_line = 0.5,
  line_alpha = 0.15,
  size_bar = 0.7,
  style = theme_lucid,
  colors = unname(social_colors(c("green", "blue"))),
  type = c("line", "dots"),
  ...
)

\method{plot}{see_performance_pp_check}(
  x,
  size_line = 0.5,
  line_alpha = 0.15,
  size_bar = 0.7,
  style = theme_lucid,
  colors = unname(social_colors(c("green", "blue"))),
  type = c("line", "dots"),
  ...
)
}
\arguments{
\item{x}{An object.}

\item{size_line}{Numeric value specifying size of line geoms.}

\item{line_alpha}{Numeric value specifying alpha of lines indicating \code{yrep}.}

\item{size_bar}{Size of bar geoms.}

\item{style}{A ggplot2-theme.}

\item{colors}{Character vector of length two, indicating the colors (in
hex-format) for points and line.}

\item{type}{Plot type for the posterior predictive checks plot. Can be \code{"line"}
(default) or \code{"dots"} (only for models with binary, integer or ordinal outcomes).}

\item{...}{Arguments passed to or from other methods.}
}
\value{
A ggplot2-object.
}
\description{
The \code{plot()} method for the \code{performance::check_predictions()} function.
}
\examples{
\dontshow{if (require("performance")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
model <- lm(Sepal.Length ~ Species * Petal.Width + Petal.Length, data = iris)
check_posterior_predictions(model)

# dot-plot style for count-models
d <- iris
d$poisson_var <- rpois(150, 1)
model <- glm(
  poisson_var ~ Species + Petal.Length + Petal.Width,
  data = d,
  family = poisson()
)
out <- check_predictions(model)
plot(out, type = "dots")
\dontshow{\}) # examplesIf}
}
