% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_okabeito.R
\name{scale_color_okabeito}
\alias{scale_color_okabeito}
\alias{scale_fill_okabeito}
\alias{scale_colour_okabeito}
\alias{scale_colour_oi}
\alias{scale_color_oi}
\alias{scale_fill_oi}
\title{Okabe-Ito color palette}
\usage{
scale_color_okabeito(
  palette = "full",
  reverse = FALSE,
  order = 1:9,
  aesthetics = "color",
  ...
)

scale_fill_okabeito(
  palette = "full",
  reverse = FALSE,
  order = 1:9,
  aesthetics = "fill",
  ...
)

scale_colour_okabeito(
  palette = "full",
  reverse = FALSE,
  order = 1:9,
  aesthetics = "color",
  ...
)

scale_colour_oi(
  palette = "full",
  reverse = FALSE,
  order = 1:9,
  aesthetics = "color",
  ...
)

scale_color_oi(
  palette = "full",
  reverse = FALSE,
  order = 1:9,
  aesthetics = "color",
  ...
)

scale_fill_oi(
  palette = "full",
  reverse = FALSE,
  order = 1:9,
  aesthetics = "fill",
  ...
)
}
\arguments{
\item{palette}{Character name of palette. Depending on the color scale, can
be one of \code{"full"}, \code{"ice"}, \code{"rainbow"}, \code{"complement"}, \code{"contrast"},
\code{"light"} (for dark themes), \code{"black_first"}, \code{full_original}, or
\code{black_first_original}. The latter three options are especially for the
Okabe-Ito color palette. The default is \code{NULL} and either \code{"contrast"} or
\code{"gradient"} is used (depending on whether \code{discrete} is \code{TRUE} or \code{FALSE}),
which are the two scale useful for discrete or gradient color scales,
respectively.}

\item{reverse}{Boolean indicating whether the palette should be reversed.}

\item{order}{A vector of numbers from 1 to 9 indicating the order of colors to use
(default: \code{1:9})}

\item{aesthetics}{A vector of names of the aesthetics that this scale
should be applied to (e.g., \code{c('color', 'fill')}).}

\item{...}{Additional arguments to pass to \code{\link[=colorRampPalette]{colorRampPalette()}}.}
}
\description{
The Okabe-Ito color palette was proposed by Okabe and Ito (2008) as a
qualitative color palette that is accessible to people with a variety of
forms of color vision deficiency. In addition to being accessible, it
includes 9 vivid colors that are readily nameable and include colors that
correspond to major primary and secondary colors (e.g., red, yellow, blue).
}
\details{
The Okabe-Ito palette is included in the base R
\code{\link[grDevices:palette]{grDevices::palette.colors()}}. These functions make this palette easier to
use with \emph{ggplot2}.

The original Okabe-Ito palette's "yellow" color is \code{"#F0E442"}. This color is
very bright and often does not show up well on white backgrounds (see
\href{https://developer.r-project.org/Blog/public/2019/11/21/a-new-palette-for-r/}{here})
for a discussion of this issue). Accordingly, by default, this function uses
a darker more "amber" color for "yellow" (\code{"#F5C710"}). This color is the
"yellow" color used in base R >4.0's \href{https://developer.r-project.org/Blog/public/2019/11/21/a-new-palette-for-r/}{default color palette}.
The palettes \code{"full"} and \code{"black_first"} use this darker yellow color. For
the original yellow color suggested by Okabe and Ito (\code{"#F0E442"}), use
palettes \code{"full_original"} or \code{"black_first_original"}.

The Okabe-Ito palette is only available as a discrete palette. For
color-accessible continuous variables, consider \link[ggplot2:scale_viridis]{the viridis palettes}.
}
\examples{
library(ggplot2)
library(see)

ggplot(iris, aes(x = Species, y = Sepal.Length, fill = Species)) +
  geom_boxplot() +
  theme_modern() +
  scale_fill_okabeito()

ggplot(iris, aes(x = Species, y = Sepal.Length, fill = Species)) +
  geom_violin() +
  theme_modern() +
  scale_fill_oi(palette = "black_first")

# for the original brighter yellow color suggested by Okabe and Ito
ggplot(iris, aes(x = Species, y = Sepal.Length, fill = Species)) +
  geom_violin() +
  theme_modern() +
  scale_fill_oi(palette = "full")

ggplot(iris, aes(x = Species, y = Sepal.Length, fill = Species)) +
  geom_violin() +
  theme_modern() +
  scale_fill_oi(order = c(1, 5, 6, 2, 4, 3, 7))
}
\references{
Okabe, M., & Ito, K. (2008). Color universal design (CUD):
How to make figures and presentations that are friendly to colorblind people.
https://jfly.uni-koeln.de/color/#pallet (Original work published 2002)
}
