% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_tol.R
\name{scale_color_tol_discrete}
\alias{scale_color_tol_discrete}
\title{Paul Tol discrete/qualitative color palettes}
\usage{
scale_color_tol_discrete(
  palette = "bright",
  reverse = FALSE,
  order = NULL,
  aesthetics = "color",
  ...
)
}
\arguments{
\item{palette}{Character name of palette. Depending on the color scale, can
be \code{"full"}, \code{"ice"}, \code{"rainbow"}, \code{"complement"},
\code{"contrast"}, \code{"light"} (for dark themes), \code{"black_first"}, \code{full_original},
or \code{black_first_original}.}

\item{reverse}{Boolean indicating whether the palette should be reversed.}

\item{aesthetics}{A vector of names of the aesthetics that this scale
should be applied to (e.g., \code{c('color', 'fill')}).}

\item{...}{Additional arguments passed to \code{discrete_scale()} when \code{discrete}
is \code{TRUE} or to \code{scale_color_gradientn()} when \code{discrete} is \code{FALSE}.}
}
\description{
Tol (2021) presents a series of palettes built with mathematical principles that
are appropriate for diverse types of data. The colors in these schemes are:
\itemize{
\item Visually distinct for all people, including viewers with color vision deficiencies
\item Distinct from black and white
\item Distinct on screen and paper,
\item Cohesive; that is, they match well together
}
}
\details{
Tol provides palettes appropriate to the 3 main types of data:
\enumerate{
\item Qualitative data – nominal or categorical data, where magnitude differences are not relevant.
\item Diverging data – data ordered between two extremes where the midpoint is important.
\item Sequential data – data ordered from low to high.
}

This function provides the qualitative palettes, as well as discrete rainbow sequential palettes.
Available palettes for each type of data are:
\itemize{
\item Qualitative: bright, high-contrast, vibrant, muted, medium-contrast, pale, dark, light, ground_cover
\item Diverging: sunset, BuRd, PRGn
\item Sequential: YlOrBr, iridescent, rainbow_discrete, rainbow_smooth
}\if{html}{\out{
<!-- For rainbow_discrete, pick the optimal set based on the number of colors,
     For rainbow_smooth, include a range argument with examples to start/end at off-white vs. purple, red vs brown
     -->
}}
\if{html}{\out{
<!-- put a table with recommended uses here -->
}}

\subsection{Colors for missing or invalid data}{

A useful feature of Tol's diverging and sequential palettes is that he
provides a recommended color to use for data that fall outside the data
range represented by the color scale (e.g., for invalid or missing data).
These colors are chosen to be highly distinct from the main color palette.
}
}
\examples{
library(ggplot2)
library(see)

ggplot(iris, aes(x = Species, y = Sepal.Length, fill = Species)) +
  geom_boxplot() +
  theme_modern() +
  scale_fill_okabeito()

ggplot(iris, aes(x = Species, y = Sepal.Length, fill = Species)) +
  geom_violin() +
  theme_modern() +
  scale_fill_oi(palette = "black_first")

# for the original brighter yellow color suggested by Okabe and Ito
ggplot(iris, aes(x = Species, y = Sepal.Length, fill = Species)) +
  geom_violin() +
  theme_modern() +
  scale_fill_oi(palette = "full")

ggplot(iris, aes(x = Species, y = Sepal.Length, fill = Species)) +
  geom_violin() +
  theme_modern() +
  scale_fill_oi(order = c(1, 5, 6, 2, 4, 3, 7))
}
\references{
Tol, P. (2021). Colour schemes (SRON/EPS Technical Note No. 09-002; Version 3.2).
SRON. https://personal.sron.nl/~pault/data/colourschemes.pdf (Original work published 2009)
}
