% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_lucid.R
\name{theme_lucid}
\alias{theme_lucid}
\title{Lucid theme}
\usage{
theme_lucid(base_size = 11, base_family = "")
}
\arguments{
\item{base_size}{base font size}

\item{base_family}{base font family}
}
\description{
A light, clear theme for ggplot.
}
\examples{
library(ggplot2)
library(see)

ggplot(iris, aes(x = Sepal.Width, y = Sepal.Length)) +
  geom_point(color = "white") +
  theme_lucid()
}
